Convert NERDTree menu to minimal to fix bug with latest version of vim (#10)
For some reason, the NERDTree menu doesn't disappear after using it
which leaves a huge buffer in the bottom that can't be closed.

There is an open pull request to fix this issue, but it hasn't been
merged yet. In the mean time, this is the temporary fix for this
situation.